Removed code used for testing & modified backend
The old patient class was removed from the front end since we will be using the backend's object from here on.  The backend was updated so that the database does not recreate itself from scratch every time create_table is called, but checks if the file exists first before creating a new database.

+++ Sprint 1/ramcare_backend.py
@@ -1,4 +1,5 @@
 import sqlite3
+from pathlib import Path
 
 #Patient object for easier editing
 class Patient:
@@ -50,21 +51,25 @@ connection_obj = sqlite3.connect('ramcare.db')
 cursor = connection_obj.cursor() 
 
 def create_table():
-    #Clear the table if it already exists
-    cursor.execute("DROP TABLE IF EXISTS RAMCARE")
+    if Path("ramcare.db").is_file():
+        print("Database already exists.")
+    else:
+        #Clear the table if it already exists
+        cursor.execute("DROP TABLE IF EXISTS RAMCARE")
 
-    #SQL works weird. You must make a query as a variable and then execute that query.
-    #You don't necessarily have to make it a variable, but I'll be doing so for now.
-    cursor.execute('''
-        CREATE TABLE RAMCARE (
-            Name VARCHAR(30) NOT NULL,
-            Blood_Type CHAR(5) NOT NULL,
-            Medicines VARCHAR(50) NOT NULL,
-            Date_of_birth CHAR(10) NOT NULL,
-            Email VARCHAR(30) NOT NULL,
-            Phone_Number CHAR(14) NOT NULL
-        );
-    ''')
+        #SQL works weird. You must make a query as a variable and then execute that query.
+        #You don't necessarily have to make it a variable, but I'll be doing so for now.
+        cursor.execute('''
+            CREATE TABLE RAMCARE (
+                Name VARCHAR(30) NOT NULL,
+                Blood_Type CHAR(5) NOT NULL,
+                Medicines VARCHAR(50) NOT NULL,
+                Date_of_birth CHAR(10) NOT NULL,
+                Email VARCHAR(30) NOT NULL,
+                Phone_Number CHAR(14) NOT NULL
+            );
+        ''')
+    
 
 def insert_data(name, blood_type, medicines, dob, email, number):
     #Insert Data
